Be defensive about the scope of Renovate.
By default Renovate will scan the repository for all `package.json` files and update those dependencies.

This might be great for some (unlikely) future Node package within this otherwise-Android repository, but the intention of this enablement is to provide dependency updates for the docs themselves and nothing else.  This is important because the docs theme and configuration (including enhancements, etc.) are provided via Node.js package.

In order to avoid Renovate operating unexpectedly on packages outside the `docs` scope (again, defensively; for the future), this adds the Renovate `packageFiles` declaration directly to the root `renovate.json` with the sole desired `package.json`: `docs/package.json`.