docs: Remove `lodash`, which doesn't seem to have ever been used.
I can't find any evidence in any docs deployment that `lodash` ever needed to
be a direct dependency.  It's possible that this is an artifact that once lived
in an early stage of the docs and just propagated out as a result of making new
docs from old docs repositories.  The `theme-example` application should help
set better precedence for this in the future.