Fix: Speed up build configuration time (#1451)
* Fix: Speed up build configuration time by lazily verifying apollo-codegen installation

Move Apollo global verification to the beginning of the task to speed up configuration process.

Before, this I/O operation (version check) would run in configuration time and take up to 5sec. This runs in every Gradle execution (even when unrelated to Apollo)

This commit fixes #1203 and moves verification to a later stage.

This means that we cannot easily fall-back to local installation that easily. That's why an exception is thrown with an appropriate message. User should disable using this feature if they don't have global installation.

* Replace println with info logs

* Move verification into exec. doFirst does not really run first