Fixed a bug related to `rootPackageName`.
`rootPackageName` was only for services before. When there are no services, the code strip the folder up to the schema.xml and set the relative path as `rootPackageName`. This caused a problems where there is an actual `rootPackageName`

Now, it is simplified. Removed `sourceFolder` in `FromFiles` since it is not necessary. Package name will be calculated by the location of the `schema.json` and `rootPackageName`