Use KSP and Moshix to generate Json adapters (#3046)
* remove the Android intelliJ workaround as IntelliJ2021 seem to be
working quite well !

* bump Kotlin and Gradle

* IntelliJ 2021.1 now supports AGP 4.1: https://youtrack.jetbrains.com/issue/IDEA-252775

* switch to MoshiX and KSP

While iterating on the compiler I had wayyyyy to many stale kapt files
issue. That should fix it.
KSP itself is still experimental but I think it's ok to use since it's
run mostly on the dev machine and the generated adapter do not do
anything crazy. On the contrary, it's a way to move the ecosystem
forward by providing feedback if needed.

* use all same versions everywhere

* add an explicit dependency to generateGrammarSource